video: tegra: bitness sanitary cleanup
* replace (u32) with (uintptr_t) which is portable;
* use portable print format strings, eg. %lx, %zd;

Reviewed-on: http://git-master/r/196623

+++ drivers/video/tegra/nvmap/nvmap_heap.h
@@ -37,7 +37,7 @@ struct nvmap_heap_block {
 
 struct nvmap_heap *nvmap_heap_create(struct device *parent, const char *name,
 				     phys_addr_t base, size_t len,
-				     unsigned int buddy_size, void *arg);
+				     size_t buddy_size, void *arg);
 
 void nvmap_heap_destroy(struct nvmap_heap *heap);
